pyproject.toml
[project]
name = "quantum-well-simulator"
version = "0.1.0"
description = "This is a Physics Research project done by Group 3"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "matplotlib",
    "mpmath",
    "numpy",
    "physipy",
    "qutip",
    "scipy",
    "sympy",
]

[project.optional-dependencies]
gpu = [
    "cupy-cuda12x",
]
